Add files via upload
So basically I have created a variable that i called numero, where the user can choose the number he would like. Then I created another one to be multiplied by the "i" number, which will encrase each time the condition is denied. Everytime this happens, the "i" is increased, with that, creating the factorial function.